move hooks file to proper nesting
