+++ src/main/content/Summary.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled'
 import { H1, H2 } from '../../SharedComponents/StyledComponents'
 
-import useGetTimeProgramming from './useGetTimeProgramming'
+import useGetTimeProgramming from '../../Hooks/useGetTimeProgramming'
 
 export default function Summary() {
   const frameworks = [
